+++ foxes/algorithms/downwind/downwind.py
@@ -988,6 +988,8 @@ class Downwind(Algorithm):
             extra_data=extra_data,
             outputs=ovars,
             parameters=calc_pars,
+            sel=sel,
+            isel=isel,
             **kwargs,
         )
         del model_data, farm_results, point_data